+++ src/imgui.lua
@@ -2,12 +2,13 @@ print("current run length", game.GameData.FullRunBiomeCount)
 
 local previousConfig = {
     biome_pool = {
-        custom_order_data = {}
     },
 }
 
+local shouldDraw
 rom.gui.add_imgui(function()
-    if rom.ImGui.Begin("Dream Dive Tweaks") then
+    shouldDraw = rom.ImGui.Begin("Dream Dive Tweaks")
+    if shouldDraw then
         DrawMenu()
         rom.ImGui.End()
     end
@@ -20,11 +21,29 @@ rom.gui.add_to_menu_bar(function()
     end
 end)
 
-local headerState = {}
-
 local scaling_applied = nil
 
+local vsync_switch = false
+
+local vsync_cache = true
+
+local function restoreVsyncOnUiClose()
+    while rom.gui.is_open() and shouldDraw do
+        game.wait(1)
+    end
+    game.SetConfigOption({Name = "VSync", Value = vsync_cache})
+    vsync_switch = false
+end
+
 function DrawMenu()
+
+    if not vsync_switch then
+        vsync_cache = game.GetConfigOptionValue({Name = "VSync"})
+        game.SetConfigOption({Name = "VSync", Value = false})
+        vsync_switch = true
+        game.thread(restoreVsyncOnUiClose)
+    end
+
     local value, selected, checked
     if rom.ImGui.CollapsingHeader("Biome pool options", rom.ImGuiTreeNodeFlags.DefaultOpen) then
         if game.CurrentHubRoom then