+++ src/scorch.lua
@@ -1,6 +1,6 @@
 modutil.mod.Path.Wrap("ApplyBurn", function (base, victim, functionArgs, triggerArgs)
     local maxStacksCopy = game.EffectData.BurnEffect.MaxStacks
-    if config.increase_scorch_cap then
+    if config.increase_scorch_cap and game.CurrentRun.IsDreamRun then
         game.EffectData.BurnEffect.MaxStacks = math.max(9999, game.EffectData.BurnEffect.MaxStacks)
     end
     base(victim, functionArgs, triggerArgs)